Crate oursh

This shell should be both POSIX compatible and yet modern and exciting. Fancy features should not be prevented by POSIX compatibility. This will effect the design of the shell.

The name of the shell is oursh which is both somewhat unique, and memorable. It's also a nice name to play with pseudo-satirical themes... right comrade? It's short (ish) and sneakily fits rs in it, which is the extension of Rust programs, the language this will be written in.

Features

POSIX compatibility w/ non-posix blocks ({@lang ... })

bash/zsh autocomplete compatibility

man / -h / --help parsing

Multi-line input

Modern scripting language (types, higher-order functions, threading?, etc)

obfuscated strings (!'password'!)

mosh like remote session support

Smart history, sync'd across devices

Package manager

Sane defaults

Fast

Usage

While this project is in early stages, there are no OS packages to use. However, you can compile and run directly from source easily.

cargo run

POSIX Reference

See the following sections for building the POSIX sh compliant program language, and interactive terminal based REPL.

32 Shell Command Language

10.2 Shell Grammar Rules

22.5 Standard I/O Streams

31.6 Built-In Utilities

31.4 Utility Description Defaults

22.3 Error Numbers

111 General Terminal Interface

22.4 Signal Concepts

Modules

job
Subprocess execution management.

program
Parsing and handling program syntax(es) of the shell.

repl
...