Module oursh::program

Parser and interpreter for the syntax(es) of the shell.

Both commands entered to the shell through STDIN and read from a file are programs, and are parsed and handled by this module.

TODO: Parse sequence of programs from stream.

POSIX Syntax

Modern Syntax

Re-exports

pub use self::basic::BasicProgram;

pub use self::posix::PosixProgram;

Modules

basic
Single command programs with no features.

posix
The ubiquitous POSIX shell command language.

Traits

Command
A command is a task given by the user as part of a Program.

Program
A program is a collection of commands given by the user.

Functions

parse
Parse a program of the given type.

parse_default
Parse a program of the default type.

Type Definitions

DefaultProgram
The default program type, used for unannotated blocks.