Module oursh::program::posix::builtin

Commands that are run from the shell directly, without forking another process.

These commands take precedence over any executables with the same name in the $PATH.

Structs

Cd

Change directory (cd) builtin.

Exit

Exit builtin, alternative to ctrl-d.

Jobs

Background job information.

Null

Noop builtin, same idea as true.

Traits

Builtin

A builtin is a custom shell command, often changing the state of the shell in some way.