Trait oursh::program::posix::builtin::Builtin

pub trait Builtin {
    fn run(
        self, 
        argv: Vec<CString>, 
        runtime: &mut Runtime<'_>
    ) -> Result<WaitStatus>;
}

A builtin is a custom shell command, often changing the state of the shell in some way.

Required methods

fn run(
self,
argv: Vec<CString>,
runtime: &mut Runtime<'_>
) -> Result<WaitStatus>

Execute the shell builtin command, returning a retult of the completion.