Trait oursh::program::Command

pub trait Command {
    fn argv(&self) -> Vec<CString>;

    fn name(&self) -> CString { ... }
}

A command is a task given by the user as part of a Program.

Each command is handled by a Job, and a single command may be run multiple times each as a new Job. Each time a command is run, the conditions within the control of the shell are reproduced; IO redirection, working directory, and even the environment are each faithfully preserved.

Required Methods

fn argv(&self) -> Vec<CString>

Return the command's arguments (including it's name).

This function returns a vector as expected by the exec(3) family of functions.

Provided Methods

fn name(&self) -> CString

Return the name of this command.

This name may not be the same as the name given to the process by the running Job.