Module oursh::repl::completion

User text completion for REPL interations.

Simple use cases for this module should be as easy as the following example taken from the current REPL.

use oursh::repl::completion::complete;

// String holding the user's input.
let mut text = "car".to_string();

// Perform the completion, on `\t` perhaps.
text = complete(&text).first();

// The user's input is updated to the complete executable.
assert_eq!("cargo", &text);

Enums

Completion

The result of a query for text completion.

Functions

complete

Return a completed (valid) program text from the partial string given.

executable_completions

Return a list of the matches from the given partial program text.

path_complete

Complete a path at the end of the given string.