Trait oursh::program::posix::builtin::Builtin

pub trait Builtin {
    fn run(argv: Vec<CString>) -> Result<WaitStatus>;
}

A builtin is a custom shell command, often changing the state of the shell in some way.

Required Methods

fn run(argv: Vec<CString>) -> Result<WaitStatus>

Execute the shell builtin command, returning a retult of the completion.