Trait oursh::program::Command

pub trait Command: Sized + Debug + Run {
    fn name(&self) -> CString { ... }
}

A command is a task given by the user as part of a Program.

Each command is handled by a Process, and a single command may be run multiple times each as a new Process. Each time a command is run, the conditions within the control of the shell are reproduced; IO redirection, working directory, and even the environment are each faithfully preserved.

Provided methods

fn name(&self) -> CString

Return the name of this command.

This name may not be the same as the name given to the process by the running Process.