Trait oursh::program::Program

pub trait Program {
    type Command: Command;
    fn parse<R: Read>(reader: R) -> Self;

    fn commands(&self) -> Vec<Self::Command>;
}

A program is a collection of commands given by the user.

Associated Types

type Command: Command

Required Methods

fn parse<R: Read>(reader: R) -> Self

Parse a whole program from the given reader.

fn commands(&self) -> Vec<Self::Command>

Return a list of all the commands in this program.